Config: use empty filesystem array for 'deny' profile
This commit fixes a copy+paste error which skipped the storage of an
empty "filesystem" array in the `deny` profile.

We also remove loose comments attached to the `deny_cfg`, `default_cfg`,
and `allow_cfg` JSON objects, as there is no guarantee that, once
serialized, they will stay where we originally wanted them to be.