macOS: prepare code for sandbox implementation
This commit isolates the sandbox implementation based on Seccomp and
Ptrace to ease the incorporation of a sandbox for macOS.

Note that, even though macOS provides a ptrace(2) implementation, that
API is far from complete. For instance, it is not possible to peek
registers, memory, or to monitor system call events from a child
process.

There aren't really too many options to monitor system calls on macOS
without requiring admin rights. The most viable approach seems to be
to implement something around Apple's Endpoint Security:

  https://developer.apple.com/documentation/endpointsecurity

Refs #16