AnonymousMemoryMap: handle MinGW64 builds
When targeting Windows, AnonymousMemoryMap is now a simple container to
host the output dataset memory area. Since the Windows port doesn't
include a sandbox yet and the OS doesn't provide a fork()-like API,
it doesn't make sense to make the implementation more complex than it
needs to be: the UDF will run on the same process as the caller's, thus
no memory copy from the mmapped address to the target buffer is needed.