find_files: fix path lookup
With the implementation move from glob() to readdir() the path lookup
test became broken. We have to concatenate the input directory with the
dentry returned from readdir() so that stat() works.

This commit also fixes the return value of that same function; it is not
considered an error if the given input directory doesn't exist. The
caller is already prepared to create it.