LuaBackend::run: fix return value Windows
On Linux and macOS, the UDF runs on a child process; when the process
fails to execute for some reason we call _exit(1) and let the parent
keep going. On Windows, the UDF runs on the same namespace as the parent
process, so we have to properly propagate the failure to the caller.

So, rather than setting `ready = false`, this commit simply lets the
function `return false`.

On a minor related change, the `waitpid()` macro now sets the `s`
argument (a pointer to the status variable) to 0 so that the compiler
doesn't complain about an unused variable.