I/O filter: load symbols from libhdf5-udf.so
Symbols required by the UDFs (such as getDataCast()) are exported by the
main libhdf5-udf library, not by libhdf5-udf-iofilter. This commit
touches many files because the variable names have also been changed to
reflect the actual library that hosts those symbols.

Note that the original code behaved correctly on Linux because of an
implicit dependency; libhdf5-udf was automatically loaded when
libhdf5-udf-iofilter was dlopened. The same doesn't happen on Windows,
though.