C++ backend: use new APIs from os.h
Use os::execCommand() as opposed to explicit calls to fork(), execvp(),
and wait4() when compiling the UDF. The same applies to
udfDatasetNames(): instead of calling pipe() + fork() + exec(), resort
to the new modified os::execCommand() API that provides the output of
the command to the caller as a std::string.

This is a first step towards having support for C/C++ UDFs on Windows.