Windows: add support for the Python backend
This commit follows the recipe of the previous changes needed to
incorporate the C/C++ and LuaJIT backends: the use of definitions from
os.h to launch new processes and the removal of glob().

The file search API that replaces glob() has been moved to a dedicated
file (file_search.[cpp,h]).