Don't take a file lock when sniffing the HDF5 file
HDF5-UDF uses a trick to overcome a limitation of the HDF5 I/O filter
API: it attempts to open the HDF5 file that hosts the UDF so it can get
a handle on that file (that way, HDF5-UDF can retrieve datasets that the
UDF depends on, for instance).

However, some applications such as HDFView take a lock on the HDF5 file
they have opened, which causes HDF5-UDF's attempt to open the same file
to fail with "Resource temporarily unavailable" (as the same lock cannot
be taken twice).

We fix this problem by temporarily setting `HDF5_USE_FILE_LOCKING=FALSE`
prior to opening the HDF5 file.