Doc: remove dependency on libhdf5-udf.so
Sphinx needs to load the UserDefinedFunction module in order to extract
its documentation. However, especially in cloud-based environments, we
don't want to have a complete installation of libhdf5-udf solely to
satisfy the automated compilation of the project docs. The workaround
we adopt is to introduce an environment variable $_DOC_BUILDER_MODE
which, if set, causes the call to dlopen() to be skipped.