Python: the class docstring needs an __init__ method
Without a declaration of an __init__() method, the introduction of a
docstring that describes the PythonLib class leads to runtime crashes
in the Python library:

  malloc(): invalid size (unsorted)

This behavior has been observed with Python 3.9.5 and 3.5.x. Oddly
enough, it only happens when HDF5-UDF is build with clang++; g++
produces an executable that runs as expected.