New backend: NVIDIA GPUDirect Storage (GDS)
This new experimental backend allows UDFs to populate dataset values
using CUDA kernels. The transfer of dataset dependencies from disk to
GPU is assisted by GPUDirect Storage: data is DMA-copied without
involving the main CPU / memory.

GDS UDFs are expected to end on '.cu'. It is assumed that the toolchain
is installed under '/usr/local/cuda'. The kernel is compiled targetting
compute_70 and compute_60. Please adjust `backend_gds.cpp` if you need
to optimize for different architectures.

A caveat of this implementation is that CuFile doesn't guarantee a
correct behavior of the library if the application calls fork(); for
this reason, this backend is not currently protected by a Sandbox.