Remove IOFILTER_READ_METADATA workaround
The IOFILTER_READ_METADATA environment variable was used to tell the
I/O filter that the reader wanted to read the JSON metadata (rather
than the actual dataset data). That was a workaround: we just wanted to
programmatically retrieve data from the first chunk, but HDF5 lacked an
API to obtain the file offset of a given chunk number.

Now that HDF5 introduced H5Dget_num_chunks() and H5Dget_chunk_info()
that limitation is gone and the implementation of the public method
HDF5_Handler::extractUDFMetadata became much simpler.