GDS: add udf_template to handle CUDA specifics
CUDA errors cannot be caught by the caller (i.e., the process that
dlopen-ed the UDF library). This seems to be a bug in CUDA+libdl,
as reported here:

  https://forums.developer.nvidia.com/t/cannot-peek-at-last-error-after-a-call-to-a-dlsym-ed-function/203143/4

Thus, we introduce a new function `hdf5_udf_last_cuda_error`, stored in
the template file, that can be used by the GDS backend to check if any
errors have been observed during the execution of the CUDA kernel. This
enables better error reporting and handling.