GDS: Use the NVIDIA GPUDirect Storage VFD driver
This is a first step towards embracing the GDS VFD work led by John
Ravi. The only limitation of the GDS VFD is that it doesn't handle
chunked data (after all, handling of compressed data belongs to a
different HDF5 layer). Consequently, we are temporarily disabling
support for decompression of Snappy-CUDA data.

The HDF5-UDF code that currently handles chunked layout is going to
be moved to a separate project for improved code reuse.